struct

+++ src/main/java/com/pfe/taxe/projetfinetude/bean/Taux.java
@@ -15,7 +15,7 @@ public class Taux {
     @OneToOne
     private CategorieLocal Categorie;
 
-    private int pourcentage;
+    private double pourcentage;
 
     public Long getId() {
         return id;
@@ -33,7 +33,7 @@ public class Taux {
         Categorie = categorie;
     }
 
-    public int getPourcentage() {
+    public double getPourcentage() {
         return pourcentage;
     }
